+++ src/main/java/mdsound/instrument/CtrQSoundInst.java
@@ -18,7 +18,7 @@ public class CtrQSoundInst extends Instrument.BaseInstrument {
     }
 
     @Override
-    public void reset(byte chipId) {
+    public void reset(int chipId) {
         device_reset_qsound(chipId);
 
         visVolume = new int[][][] {
@@ -28,22 +28,22 @@ public class CtrQSoundInst extends Instrument.BaseInstrument {
     }
 
     @Override
-    public int start(byte chipId, int clock) {
+    public int start(int chipId, int clock) {
         return device_start_qsound(chipId, 4000000);
     }
 
     @Override
-    public int start(byte chipId, int clock, int clockValue, Object... option) {
+    public int start(int chipId, int clock, int clockValue, Object... option) {
         return device_start_qsound(chipId, clockValue);
     }
 
     @Override
-    public void stop(byte chipId) {
+    public void stop(int chipId) {
         device_stop_qsound(chipId);
     }
 
     @Override
-    public void update(byte chipId, int[][] outputs, int samples) {
+    public void update(int chipId, int[][] outputs, int samples) {
         qsound_update(chipId, outputs, samples);
 
         visVolume[chipId][0][0] = outputs[0][0];
@@ -51,98 +51,98 @@ public class CtrQSoundInst extends Instrument.BaseInstrument {
     }
 
     @Override
-    public int write(byte chipId, int port, int adr, int data) {
-        qsound_w(chipId, adr, (byte) data);
+    public int write(int chipId, int port, int adr, int data) {
+        qsound_w(chipId, adr, data);
         return 0;
     }
 
-    private int device_start_qsound(byte chipId, int clock) {
+    private int device_start_qsound(int chipId, int clock) {
         CtrQsound chip = QSoundData[chipId];
         return chip.start2(clock);
     }
 
-    private void device_stop_qsound(byte chipId) {
+    private void device_stop_qsound(int chipId) {
         device_stop_qsound_ctr(chipId);
     }
 
-    private void device_reset_qsound(byte chipId) {
+    private void device_reset_qsound(int chipId) {
         device_reset_qsound_ctr(chipId);
         // need to wait until the chips is ready before we start writing to it ...
         // we do this by time travel.
         qsoundc_wait_busy(chipId);
     }
 
-    public void qsound_w(byte chipId, int offset, byte data) {
+    public void qsound_w(int chipId, int offset, int data) {
         CtrQsound chip = QSoundData[chipId];
         chip.write2(offset, data);
     }
 
-    private void qsound_update(byte chipId, int[][] outputs, int samples) {
+    private void qsound_update(int chipId, int[][] outputs, int samples) {
         qsoundc_update(chipId, outputs, samples);
     }
 
-    public void qsound_write_rom(byte chipId, int romSize, int dataStart, int dataLength, byte[] romData) {
+    public void qsound_write_rom(int chipId, int romSize, int dataStart, int dataLength, byte[] romData) {
         qsoundc_write_rom(chipId, romSize, dataStart, dataLength, romData, 0);
     }
 
-    public void qsound_write_rom(byte chipId, int romSize, int dataStart, int dataLength, byte[] romData, int srcStartAddress) {
+    public void qsound_write_rom(int chipId, int romSize, int dataStart, int dataLength, byte[] romData, int srcStartAddress) {
         qsoundc_write_rom(chipId, romSize, dataStart, dataLength, romData, srcStartAddress);
     }
 
-    public void qsound_set_mute_mask(byte chipId, int muteMask) {
+    public void qsound_set_mute_mask(int chipId, int muteMask) {
         qsoundc_set_mute_mask(chipId, muteMask);
     }
 
     private static final int MAX_CHIPS = 0x02;
     CtrQsound[] QSoundData = new CtrQsound[MAX_CHIPS];
 
-    private int device_start_qsound_ctr(byte chipId, int clock) {
+    private int device_start_qsound_ctr(int chipId, int clock) {
         QSoundData[chipId] = new CtrQsound();
         CtrQsound chip = QSoundData[chipId];
         return chip.start(clock);
     }
 
-    private void device_stop_qsound_ctr(byte chipId) {
+    private void device_stop_qsound_ctr(int chipId) {
         CtrQsound chip = QSoundData[chipId];
     }
 
-    private void device_reset_qsound_ctr(byte chipId) {
+    private void device_reset_qsound_ctr(int chipId) {
         CtrQsound chip = QSoundData[chipId];
         chip.reset();
     }
 
-    private byte qsoundc_r(byte chipId, int offset) {
+    private byte qsoundc_r(int chipId, int offset) {
         CtrQsound chip = QSoundData[chipId];
         return chip.read(offset);
     }
 
-    private void qsoundc_w(byte chipId, int offset, byte data) {
+    private void qsoundc_w(int chipId, int offset, byte data) {
         CtrQsound chip = QSoundData[chipId];
         chip.write(offset, data);
     }
 
-    private void qsoundc_write_data(byte chipId, byte address, int data) {
+    private void qsoundc_write_data(int chipId, byte address, int data) {
         CtrQsound chip = QSoundData[chipId];
         chip.writeData(address, data);
     }
 
-    private void qsoundc_update(byte chipId, int[][] outputs, int samples) {
+    private void qsoundc_update(int chipId, int[][] outputs, int samples) {
         CtrQsound chip = QSoundData[chipId];
         chip.update(outputs, samples);
     }
 
-    private void qsoundc_write_rom(byte chipId, int romSize, int dataStart, int dataLength, byte[] romData, int srcStartAddress) {
+    private void qsoundc_write_rom(int chipId, int romSize, int dataStart, int dataLength, byte[] romData, int srcStartAddress) {
         CtrQsound chip = QSoundData[chipId];
         chip.writeRom(romSize, dataStart, dataLength, romData, srcStartAddress);
     }
 
-    private void qsoundc_set_mute_mask(byte chipId, int muteMask) {
+    private void qsoundc_set_mute_mask(int chipId, int muteMask) {
         CtrQsound chip = QSoundData[chipId];
         if (chip == null) return;
         chip.setMuteMask(muteMask);
     }
 
-    private void qsoundc_wait_busy(byte chipId) {
+    private void qsoundc_wait_busy(int chipId) {
         CtrQsound chip = QSoundData[chipId];
         chip.waitBusy();
     }
