+++ src/main/java/mdsound/instrument/P86Inst.java
@@ -19,38 +19,38 @@ public class P86Inst extends Instrument.BaseInstrument {
     }
 
     @Override
-    public void reset(byte chipId) {
+    public void reset(int chipId) {
         info.init();
     }
 
     @Override
-    public int start(byte chipId, int clock) {
+    public int start(int chipId, int clock) {
         return start(chipId, clock, 0);
     }
 
     @Override
-    public int start(byte chipId, int clock, int clockValue, Object... option) {
+    public int start(int chipId, int clock, int clockValue, Object... option) {
         info.samplingRate = clock;
         reset(chipId);
         return clock;
     }
 
     @Override
-    public void stop(byte chipId) {
+    public void stop(int chipId) {
         // none
     }
 
     @Override
-    public void update(byte chipId, int[][] outputs, int samples) {
+    public void update(int chipId, int[][] outputs, int samples) {
         info.update(outputs, samples);
     }
 
     @Override
-    public int write(byte chipId, int port, int adr, int data) {
+    public int write(int chipId, int port, int adr, int data) {
         return info.write(port, adr, data);
     }
 
-    public int loadPcm(byte chipId, byte address, byte data, byte[] pcmData) {
+    public int loadPcm(int chipId, byte address, byte data, byte[] pcmData) {
         return info.loadPcm(0, address, data, pcmData);
     }
 }
